# openapi.yaml
openapi: 3.0.3
info:
  title: BookedIRL Backend API
  version: 0.1.1
servers:
  - url: http://localhost:4000
  - url: https://bookedirlbackend.onrender.com
...
paths:
...
  /services:
    get:
      tags: [services]
      security:
        - bearerAuth: []
      summary: List services (optionally by userIds)
      parameters:
        - name: userIds
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      tags: [services]
      security:
        - bearerAuth: []
      summary: Create service for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
  /services/{serviceId}:
    get:
      tags: [services]
      security:
        - bearerAuth: []
      summary: Get service by id
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    patch:
      tags: [services]
      security:
        - bearerAuth: []
      summary: Update service by id
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    delete:
      tags: [services]
      security:
        - bearerAuth: []
      summary: Delete service by id
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /users/{userId}/services:
    get:
      tags: [services]
      security:
        - bearerAuth: []
      summary: List services for user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      tags: [services]
      security:
        - bearerAuth: []
      summary: Create service for user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    IdResponse:
      type: object
      properties:
        id:
          type: string
      required: [id]
...
    Service:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        price_dollars:
          type: integer
        duration_minutes:
          type: integer
        is_active:
          type: boolean
      required: [id, user_id, title, price_dollars, duration_minutes, is_active]
    ServiceCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priceDollars:
          type: integer
        durationMinutes:
          type: integer
      required: [title, priceDollars, durationMinutes]
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        priceDollars:
          type: integer
        durationMinutes:
          type: integer
        isActive:
          type: boolean
      additionalProperties: false
...